moved download tools tests to its correct test file
